requirements.txt
requests
python-dotenv
pandas 
openpyxl